BVH & AABB implemented on the GPU
Implementing the BVH & AABB code on the GPU brought massive speedups, so
much so that scenes that would kill my GPU before can now run.

Speaking of killing GPUs, there are still some when the scene to its limits.
samples_per_pixel seems especially heavy (obviously max_samples is also
heavy, but samples_per_pixel especially so). Maybe this performance is just the con
of using a M2 for graphics processing instead of an actual graphics
card.

Another issue is the time to first frame. This is because of the BVH
wrangling to get in into the right shape for the uniform buffer, although
I'm not sure how much of a speedup could be expected here.